// src/app/types/component-types-sb.d.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import { StoryblokRichTextNodeTypes } from "@storyblok/richtext";
import type { ISbStoryData } from "storyblok";
export interface FeatureStoryblok {
  headline: string;
  content: string;
  component: "feature";
  _uid: string;
  [k: string]: any;
}

export interface GridStoryblok {
  headline: string;
  columns: (FeatureStoryblok | TestimonialStoryblok)[];
  component: "grid";
  _uid: string;
  [k: string]: any;
}

export interface HeroStoryblok {
  headline: string;
  content: string;
  component: "hero";
  _uid: string;
  [k: string]: any;
}

export interface PageStoryblok {
  blocks?: (FeatureStoryblok | GridStoryblok | HeroStoryblok | RecommendedToursStoryblok | TestimonialStoryblok)[];
  component: "page";
  _uid: string;
  [k: string]: any;
}

export interface RecommendedToursStoryblok {
  headline: string;
  tours: (ISbStoryData<TourStoryblok> | string)[];
  component: "recommended_tours";
  _uid: string;
  [k: string]: any;
}

export interface TestimonialStoryblok {
  name: string;
  comment: string;
  component: "testimonial";
  _uid: string;
  [k: string]: any;
}

export interface AssetStoryblok {
  alt: string | null;
  copyright?: string | null;
  fieldtype: "asset";
  id: number;
  filename: string | null;
  name: string;
  title: string | null;
  focus: string | null;
  meta_data?: {
    [k: string]: any;
  };
  source?: string | null;
  is_external_url?: boolean;
  is_private?: boolean;
  src?: string;
  updated_at?: string;
  width?: number | null;
  height?: number | null;
  aspect_ratio?: number | null;
  public_id?: string | null;
  content_type?: string;
  [k: string]: any;
}

export interface StoryblokRichTextNode<T = string> {
  type: StoryblokRichTextNodeTypes;
  content: StoryblokRichTextNode<T>[];
  attrs?: Record<string, any>;
  text?: string;
  [k: string]: any;
}

export interface TourStoryblok {
  name: string;
  introduction: string;
  price: string;
  main_image: AssetStoryblok;
  location: "" | "Taipei" | "Jiufen" | "Kaohsiung";
  body: StoryblokRichTextNode;
  component: "tour";
  _uid: string;
  [k: string]: any;
}